Add support for spawn start method (DM-27118)
Pipetask has new option --start-method which tkaes the name of the start
methof from multiprocessing (spawn, fork, forkserver). Default method is
now set as "spawn" on MacOS and "fork" on Linux. Some care was needed
when unpickling dimensions, this has to be done after unpickling
registry. Unpickling is ordered manually now. As usual logging causes
issues, with spawn it needs to be re-initialized in every child, added
support for recording and replaying configuration (part of that is in
`daf_butler`).