Improve typing and docs for collection sequences.
DatasetTypeRestriction hasn't been used here for a while, so those
annotations and docs are just wrong.  Docs about the inputs attribute
being updated to include outputs were wrong, too, and because it isn't
being updated, I've made it a tuple instead of a list so there's no
doubt about that.